people saga moved to personList
